refactor: view pages and serparating how the views should be strutured
