+++ src/frontend/src/views/home/jobs/JobAssignment.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import NavigationBar from '../components/home/NavigationBar.vue'
+import NavigationBar from '../../../components/home/NavigationBar.vue'
 </script>
 
 <template>
